Broke WEBSOCKETS module into pieces too.
