+++ websocket/src/main/java/org/nanohttpd/protocols/websockets/State.java
@@ -1,0 +1,9 @@
+package org.nanohttpd.protocols.websockets;
+
+public enum State {
+    UNCONNECTED,
+    CONNECTING,
+    OPEN,
+    CLOSING,
+    CLOSED
+}
